Update cucumber to 2.3.1 with tests

+++ test/features/support/env.js
@@ -14,46 +14,110 @@
  * governing permissions and limitations under the License.
  **/
 
-var configure = function () {
+const
+    {defineSupportCode} = require('cucumber');
 
-    this.setDefaultTimeout(30 * 1000);
+defineSupportCode(function ({After, Before, setDefaultTimeout, defineParameterType}) {
 
-    this.After(function () {
+    defineParameterType({
+        regexp:      /\d+(?:st|nd|rd|th)/,
+        transformer: function (string) {
+            return parseInt(string.replace(/st|nd|rd|th/, ''));
+        },
+        typeName:    'ordinal'
+    });
+
+    setDefaultTimeout(5 * 1000);
+
+    After(function () {
 
         browser.clearMockModules();
 
-        browser.executeScript(function () {
-            try {
+        return browser.executeScript(function () {
+            if (window.localStorageSupport) {
                 window.localStorage.clear();
-            } catch (e) {
             }
         });
-
     });
 
-    this.Before(function () {
+    Before(function () {
 
-        if (browser.browserName) {
+        const world = this;
+
+        if (world.browserName) {
             return;
         }
 
         return browser
             .getCapabilities()
             .then(function (capabilities) {
-                browser.browserName = capabilities.get('browserName');
+                world.browserName = capabilities.get('browserName');
             });
     });
 
-    // set default config to config.json in fixtures directory
-    this.Before(function () {
+    Before(function () {
 
         return browser.addMockModule('app', function () {
+
             window.addToHomescreen = angular.noop;
-            angular.module('app').run(function () {
+
+            angular.module('app').run(function ($rootScope) {
                 window.configLocation = './fixtures/config/default.json';
+
+                $rootScope.$on('$stateChangeSuccess', function () {
+                    window.$stateIsResolved = true;
+                });
             });
         });
     });
-};
 
-module.exports = configure;
+    Before(function () {
+
+        return browser.addMockModule('ngMockModule', function () {
+
+            angular.module('ngMockModule', ['ngMockE2E']).run(function ($httpBackend) {
+
+                let mockFeeds = ['lrYLc95e', 'WXu7kuaW', 'Q352cyuc', 'oR7ahO0J'];
+
+                // mock each feed request defined in mockFeeds
+                angular.forEach(mockFeeds, function (id) {
+
+                    $httpBackend.whenGET('https://content.jwplatform.com/v2/playlists/' + id)
+                        .respond(function () {
+                            var request = new XMLHttpRequest();
+
+                            request.open('GET', './fixtures/feed/' + id + '.json', false);
+                            request.send(null);
+
+                            return [request.status, request.response, {}];
+                        });
+                });
+
+                // let all other requests pass
+                $httpBackend.whenGET(/\S+/)
+                    .passThrough();
+            });
+        });
+    });
+
+    Before(function () {
+
+        return browser.addMockModule('disableAnimate', function () {
+
+            angular.module('disableAnimate', []).run(function ($animate) {
+                var style       = document.createElement('style');
+                style.type      = 'text/css';
+                style.innerHTML = '* {' +
+                    '-webkit-transition: none !important;' +
+                    '-moz-transition: none !important;' +
+                    '-o-transition: none !important;' +
+                    '-ms-transition: none !important;' +
+                    'transition: none !important;' +
+                    '}';
+                document.getElementsByTagName('head')[0].appendChild(style);
+
+                $animate.enabled(false);
+            });
+        });
+    });
+});
